+++ CivilWar/CivilWar Server/Soldier.h
@@ -9,15 +9,15 @@ public:
 	~Soldier();
 
 	Point getCenter() const { return m_center; }
-	int getX() const { return m_center.x; }
-	int getY() const { return m_center.y; }
+	float getX() const { return m_center.x; }
+	float getY() const { return m_center.y; }
 	Radius getRadius() const { return m_radius; }
 	Health getHealth() const { return m_health; }
 	Speed getSpeed() const { return m_speed; }
 
 	void setCenter(Point center) { m_center = center; }
-	void setX(int x) { m_center.x = x; }
-	void setY(int y) { m_center.y = y; }
+	void setX(float x) { m_center.x = x; }
+	void setY(float y) { m_center.y = y; }
 	void setRadius(Radius radius) { m_radius = radius; }
 	void setHealth(Health health) { m_health = health; }
 	void setHealth(Speed speed) { m_speed = speed; }
